refactor: share task content parser
